+++ soya/src/browser/ComponentBrowserSegment.js
@@ -1,0 +1,8 @@
+import LocalSegment from 'soya/lib/data/redux/segment/local/LocalSegment';
+
+/**
+ * @CLIENT_SERVER
+ */
+export default class ComponentBrowserSegment extends LocalSegment {
+  
+}