+++ soya/src/browser/ComponentList.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+/**
+ *
+ */
+export default class ComponentList extends React.Component {
+
+}