+++ test/integration/src/components/common/PhotoCaption/PhotoCaptionDoc.js
@@ -1,0 +1,26 @@
+import React from 'react';
+
+/**
+ * @CLIENT_SERVER
+ */
+export default class PhotoCaptionDoc extends React.Component {
+  render() {
+    return <div>
+      <h3>Sample Documentation</h3>
+      <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Ut fringilla nulla id elit maximus, ut tincidunt velit cursus. Sed molestie a erat porttitor volutpat. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Proin volutpat turpis nec velit faucibus, a mattis urna ultrices. Sed sagittis ut orci vel sollicitudin. Praesent lobortis elementum pulvinar.</p>
+      <ul>
+        <li>Test short list.</li>
+        <li>Test list with medium length. With <code>code</code>.</li>
+        <li>Curabitur non felis ac justo dignissim lacinia non in urna. Sed elementum lorem id urna molestie hendrerit. Quisque congue congue convallis. Quisque scelerisque neque vitae efficitur pellentesque. Mauris pulvinar neque lectus, eu pharetra lectus viverra ut. Cras auctor leo tellus, quis egestas lectus placerat in.</li>
+      </ul>
+      <p>Short paragraph is sometimes needed. Also <strong>strong sentences</strong> and <em>minor emphasis</em>.</p>
+      <ol>
+        <li>Test short list.</li>
+        <li>Test list with medium length. With <code>code</code>.</li>
+        <li>Curabitur non felis ac justo dignissim lacinia non in urna. Sed elementum lorem id urna molestie hendrerit. Quisque congue congue convallis. Quisque scelerisque neque vitae efficitur pellentesque. Mauris pulvinar neque lectus, eu pharetra lectus viverra ut. Cras auctor leo tellus, quis egestas lectus placerat in.</li>
+      </ol>
+      <p>If further demonstration of code is needed:</p>
+      <pre>import React from 'react'</pre>
+    </div>;
+  }
+}