+++ test/integration/src/components/common/PhotoCaption/PhotoCaptionThumbnail.js
@@ -3,6 +3,9 @@ import PhotoCaption from './PhotoCaption.js';
 
 import url from './mountains.jpg';
 
+/**
+ * @CLIENT_SERVER
+ */
 export default class PhotoCaptionThumbnail extends React.Component {
   render() {
     return <div>